tidy up server routes into seperate files
