+++ client/src/api/index.js
@@ -21,7 +21,7 @@ export function getHistory() {
 
 export function getSchedules() {
 
-  return fetch(`${SERVER_URL}/api/todaySchedules`)
+  return fetch(`${SERVER_URL}/api/schedules/today`)
   .then(res => res.json())
   .then(res => res.schedules)
 }
